@@ -19,16 +19,16 @@
         "postcss": "^8.4.18",
         "prettier": "^3.4.2",
         "svelte-dnd-action": "^0.9.40",
-        "svelte-preprocess": "^5.1.2",
+        "svelte-preprocess": "^6.0.0",
         "tailwindcss": "^3.2.1",
         "ts-loader": "^9.5.1",
-        "typescript": "^5.3.3",
-        "vite": "^4.0.0",
+        "typescript": "^5.5.0",
+        "vite": "^5.4.4",
         "vue": "^3.2.41"
     },
     "dependencies": {
         "@inertiajs/svelte": "^1.0.14",
-        "@sveltejs/vite-plugin-svelte": "^2.4.6",
+        "@sveltejs/vite-plugin-svelte": "^4.0.0",
         "@tailwindcss/typography": "^0.5.10",
         "chart.js": "^4.4.1",
         "class-variance-authority": "^0.7.0",
