@@ -2,16 +2,30 @@
     import BaseButton from "./BaseButton.svelte";
     import type { ButtonSize, ButtonShape, ButtonVariant } from "$types";
     
-    export let href: string | undefined = undefined;
-    export let loading: boolean = false;
-    export let size: ButtonSize = "md";
-    export let variant: ButtonVariant = "danger";
-    export let shape: ButtonShape = undefined;
+    interface Props {
+        href?: string | undefined;
+        loading?: boolean;
+        size?: ButtonSize;
+        variant?: ButtonVariant;
+        shape?: ButtonShape;
+        children?: import('svelte').Snippet;
+        [key: string]: any
+    }
+
+    let {
+        href = undefined,
+        loading = false,
+        size = "md",
+        variant = "danger",
+        shape = undefined,
+        children,
+        ...rest
+    }: Props = $props();
 </script>
 
 <BaseButton {size} {shape} {variant} {loading} {href}
     on:click
-    {...$$restProps}
+    {...rest}
 >
-    <slot />
+    {@render children?.()}
 </BaseButton>
