@@ -1,23 +1,38 @@
 <script lang="ts">
+    import { createBubbler } from 'svelte/legacy';
+
+    const bubble = createBubbler();
     import { containerColors as colors } from "$lib/config";
     import { twMerge } from "tailwind-merge";
 
-    export let color: "blue" | "red" | "green" | "yellow" | "gray" | "indigo" | "purple" | "pink" | "orange" | "teal" | "cyan" | "white" | "black" | "ghost"
-        | "primary" | "secondary" | "accent" | "neutral" | "info" | "success" | "warning" | "error" | "base-100" | "base-200" = "primary";
-    export let type: "button" | "submit" | "reset" = "button";
-    export let href: string = "";
+    interface Props {
+        color?: "blue" | "red" | "green" | "yellow" | "gray" | "indigo" | "purple" | "pink" | "orange" | "teal" | "cyan" | "white" | "black" | "ghost"
+        | "primary" | "secondary" | "accent" | "neutral" | "info" | "success" | "warning" | "error" | "base-100" | "base-200";
+        type?: "button" | "submit" | "reset";
+        href?: string;
+        children?: import('svelte').Snippet;
+        [key: string]: any
+    }
 
-    const css = twMerge("border rounded px-2 max-h-6 bg-slate-300 text-black capitalize",colors[color], $$restProps.class);
-    const {class: _, ...rest} = $$restProps;
+    let {
+        color = "primary",
+        type = "button",
+        href = "",
+        children,
+        ...rest_1
+    }: Props = $props();
+
+    const css = twMerge("border rounded px-2 max-h-6 bg-slate-300 text-black capitalize",colors[color], rest_1.class);
+    const {class: _, ...rest} = rest_1;
 
 </script>
 
 {#if !href}
-<button {type} class={css} on:click {...rest}>
-    <slot></slot>
+<button {type} class={css} onclick={bubble('click')} {...rest}>
+    {@render children?.()}
 </button>
 {:else}
-<a role="button" {href} class={css} on:click {...rest}>
-        <slot></slot>
+<a role="button" {href} class={css} onclick={bubble('click')} {...rest}>
+        {@render children?.()}
 </a>
 {/if}
