@@ -1,8 +1,13 @@
 <script lang="ts">
     import { createEventDispatcher } from "svelte";
 
-    export let checked: boolean = false;
-    export let value: any = null;
+    interface Props {
+        checked?: boolean;
+        value?: any;
+        [key: string]: any
+    }
+
+    let { checked = $bindable(false), value = $bindable(null), ...rest }: Props = $props();
 
     const dispatch = createEventDispatcher();
 </script>
@@ -13,6 +18,6 @@
         bind:value
         bind:checked
         class="rounded dark:bg-gray-900 border-gray-300 dark:border-gray-700 text-indigo-600 shadow-sm focus:ring-indigo-500 dark:focus:ring-indigo-600 dark:focus:ring-offset-gray-800"
-        {...$$restProps}
+        {...rest}
         />
 
