@@ -1,20 +1,28 @@
 <script lang="ts">
+    import { createBubbler } from 'svelte/legacy';
+
+    const bubble = createBubbler();
     import { twMerge } from 'tailwind-merge';
 
-    export let value: string | undefined;
+    interface Props {
+        value: string | undefined;
+        [key: string]: any
+    }
+
+    let { value = $bindable(), ...rest }: Props = $props();
 
 
     const cssClass = twMerge(
         "input input-bordered w-full max-w-xs",
-        $$restProps.class
+        rest.class
     );
 
 </script>
 
 <input
     bind:value
-    on:input on:change on:blur on:keydown on:keyup on:keypress on:focus
-    {...$$restProps}
+    oninput={bubble('input')} onchange={bubble('change')} onblur={bubble('blur')} onkeydown={bubble('keydown')} onkeyup={bubble('keyup')} onkeypress={bubble('keypress')} onfocus={bubble('focus')}
+    {...rest}
     type="date"
     class={cssClass}
 />
