@@ -1,7 +1,13 @@
 <script lang="ts">
     import { twMerge } from 'tailwind-merge';
+    interface Props {
+        children?: import('svelte').Snippet;
+        [key: string]: any
+    }
+
+    let { children, ...rest }: Props = $props();
 </script>
 
-<fieldset {...$$restProps} class={twMerge("border rounded p-5",$$restProps.class)}>
-    <slot />
+<fieldset {...rest} class={twMerge("border rounded p-5",rest.class)}>
+    {@render children?.()}
 </fieldset>
