@@ -1,9 +1,14 @@
 <script lang="ts">
-    export let message: string | undefined;
+    interface Props {
+        message: string | undefined;
+        [key: string]: any
+    }
+
+    let { message, ...rest }: Props = $props();
 </script>
 
 {#if message}
-    <div {...$$restProps}>
+    <div {...rest}>
         <p class="text-sm text-red-600 dark:text-red-400">
             { message }
         </p>
