@@ -1,10 +1,26 @@
 <script lang="ts">
+    import { createBubbler } from 'svelte/legacy';
+
+    const bubble = createBubbler();
     import { twMerge } from 'tailwind-merge';
 
-    export let value: string | number | null = null;
-    export let items: {label:string,value: number|string}[]  = [];
-    export let name: string;
-    export let mapping: { labelColumn: string, valueColumn: string } = { labelColumn: "label", valueColumn: "value" };
+    interface Props {
+        value?: string | number | null;
+        items?: {label:string,value: number|string}[];
+        name: string;
+        mapping?: { labelColumn: string, valueColumn: string };
+        children?: import('svelte').Snippet;
+        [key: string]: any
+    }
+
+    let {
+        value = $bindable(null),
+        items = [],
+        name,
+        mapping = { labelColumn: "label", valueColumn: "value" },
+        children,
+        ...rest
+    }: Props = $props();
 
 
     let options: {label: string, value: string | number}[] = items?.map((item: any) => {
@@ -18,18 +34,18 @@
 
 
 
-<select bind:value {name} {...$$restProps}  class="select select-bordered w-full max-w-xs"
-    on:change
-    on:click
-    on:input
-    on:blur
-    on:focus
+<select bind:value {name} {...rest}  class="select select-bordered w-full max-w-xs"
+    onchange={bubble('change')}
+    onclick={bubble('click')}
+    oninput={bubble('input')}
+    onblur={bubble('blur')}
+    onfocus={bubble('focus')}
 >
     {#if options?.length > 0}
         {#each options ?? [] as item, index}
             <option value={item.value}>{item.label}</option>
         {/each}
     {:else}
-        <slot />
+        {@render children?.()}
     {/if}
 </select>
