@@ -1,12 +1,18 @@
 <script lang="ts">
-  export let title: string | undefined = undefined;
+  interface Props {
+    title?: string | undefined;
+    children?: import('svelte').Snippet;
+    [key: string]: any
+  }
+
+  let { title = undefined, children, ...rest }: Props = $props();
   
-  let modal: HTMLDialogElement;
+  let modal: HTMLDialogElement = $state();
 </script>
 
 
 
-<dialog class="modal overflow-clip" bind:this={modal} {...$$restProps}>
+<dialog class="modal overflow-clip" bind:this={modal} {...rest}>
 
     <div class="modal-box">
         {#if title}
@@ -17,7 +23,7 @@
         <p class="py-4">Press ESC key or click the button below to close</p>
         <div class="modal-action">
           <form method="dialog" class="flex flex-col w-full gap-3 items-between">
-            <slot />
+            {@render children?.()}
             <button class="btn">Close</button>
           </form>
         </div>
