@@ -1,6 +1,11 @@
 <script lang="ts">
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
 
-    let div: HTMLDivElement;
+    let { children }: Props = $props();
+
+    let div: HTMLDivElement = $state();
     const defaultClasses = `border overflow-hidden shadow-md 
     sm:rounded-lg p-3 flex items-center justify-center hover:cursor-grab transition-all duration-300 ease-in-out`;
 
@@ -10,5 +15,5 @@
     role="contentinfo"
 >
     
-    <slot/>
+    {@render children?.()}
 </div>
