@@ -8,6 +8,12 @@
     import route from "$vendor/tightenco/ziggy";
     import { fade, fly, slide } from "svelte/transition";
     import { sineInOut } from "svelte/easing";
+    interface Props {
+        header?: import('svelte').Snippet;
+        children?: import('svelte').Snippet;
+    }
+
+    let { header, children }: Props = $props();
 
     const { auth, roles } = $page.props;
 
@@ -46,10 +52,10 @@
         <Navbar {menu}/>
 
         <!-- Page Heading -->
-        {#if $$slots.header}
+        {#if header}
             <header class="bg-white dark:bg-gray-800 shadow">
                 <div class="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8">
-                    <slot name="header" />
+                    {@render header?.()}
                 </div>
             </header>
         {/if}
@@ -57,7 +63,7 @@
         <!-- Page Content -->
         <main class="sm:p-6 py-2"  transition:fade={{duration: 400, easing: sineInOut}}>
             <MiniToast show={false} />
-            <slot />
+            {@render children?.()}
         </main>
     </div>
     <footer class="w-screen flex items-center justify-center h-12">
