@@ -1,7 +1,12 @@
-<script setup>
+<script lang="ts" setup>
     import PageTitle from '$components/UI/PageTitle.svelte';
 import ApplicationLogo from '../Components/UI/ApplicationLogo.svelte';
 import { Link } from '@inertiajs/svelte';
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
 </script>
     
 <div class="min-h-screen flex flex-col sm:justify-center items-center pt-6 sm:pt-0 bg-gray-100 dark:bg-gray-900">
@@ -14,6 +19,6 @@ import { Link } from '@inertiajs/svelte';
     <div
         class="w-full sm:max-w-md mt-6 px-6 py-4 bg-white dark:bg-gray-800 shadow-md overflow-hidden sm:rounded-lg"
     >
-        <slot />
+        {@render children?.()}
     </div>
 </div>
