@@ -1,4 +1,7 @@
 <script lang="ts">
+    import { createBubbler, preventDefault } from 'svelte/legacy';
+
+    const bubble = createBubbler();
     import Clock from '$components/App/TimeLog/Clock.svelte';
     import DashboardItem from '$components/UI/DashboardItem.svelte';
     import ClockingForm from '$components/App/TimeLog/ClockingForm.svelte';
@@ -8,11 +11,15 @@
     import PrimaryButton from '$components/Buttons/PrimaryButton.svelte';
     import NewLogModal from '$components/Modals/NewLogModal.svelte';
 
-    export let auth: any = null;
+    interface Props {
+        auth?: any;
+    }
+
+    let { auth = null }: Props = $props();
 
     const dropzones = writable(Array(4).fill(null));
 
-    let newLogModalOpen = false;
+    let newLogModalOpen = $state(false);
 
 </script>
 
@@ -20,14 +27,16 @@
     <h2 class="font-semibold text-xl text-gray-800 dark:text-gray-200 leading-tight">Dashboard - {auth.user.name}</h2>
 
 
-    <div class="py-12" on:dragover|preventDefault role="application" >
+    <div class="py-12" ondragover={preventDefault(bubble('dragover'))} role="application" >
         <div class="max-w-7xl mx-auto sm:px-6 lg:px-8 grid gap-5 sm:grid-cols-2 grid-cols-1">
             <DashboardItem>
                 <Clock />
             </DashboardItem>
             <DashboardItem>
                 <ClockingForm>
-                    <ElapsedMarker slot="indicator"/>
+                    {#snippet indicator()}
+                                        <ElapsedMarker />
+                                    {/snippet}
                 </ClockingForm>
             </DashboardItem>
             <DashboardItem>
