@@ -1,13 +1,25 @@
 <script lang="ts">
+    import { run } from 'svelte/legacy';
+
     import { Chart } from 'chart.js/auto';
     
-    export let budget: number = 0;
-    export let spent: number = 0;
-    export let title: string;
-    export let width: number | string = 600;
-    export let height: number | string = 600;
+    interface Props {
+        budget?: number;
+        spent?: number;
+        title: string;
+        width?: number | string;
+        height?: number | string;
+    }
 
-    let canvas: HTMLCanvasElement;
+    let {
+        budget = 0,
+        spent = 0,
+        title,
+        width = 600,
+        height = 600
+    }: Props = $props();
+
+    let canvas: HTMLCanvasElement = $state();
 
 
     const data = {
@@ -17,12 +29,14 @@
         }]
     };
 
-    $: if(canvas) {
-        const chart = new Chart(canvas, {
-            type: 'pie',
-            data: data,
-        });
-    }
+    run(() => {
+        if(canvas) {
+            const chart = new Chart(canvas, {
+                type: 'pie',
+                data: data,
+            });
+        }
+    });
 </script>
 
 <div class="modal-action flex justify-center h-96">
