@@ -2,7 +2,7 @@
     import { Calendar } from "$lib/components/ui/calendar";
     import { getLocalTimeZone, today } from "@internationalized/date";
    
-    let value = today(getLocalTimeZone());
+    let value = $state(today(getLocalTimeZone()));
   </script>
    
   <Calendar bind:value class="rounded-md border shadow" />
